# specs/001-feat-pamali-smart-campus-ops-hub/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Campus Operations Hub API
  version: 1.0.0
  description: Contract-first API for auth, facilities, bookings, maintenance, notifications, and analytics.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /bookings:
    get:
      summary: Get pending bookings for current user's approval role
      description: Returns bookings pending approval based on approver's role (LECTURER, ADMIN, FACILITY_MANAGER).
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of pending bookings for approval
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookingResponse"
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  number:
                    type: integer
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create booking request
      description: Applies quota, role policy, overlap checks, and approval workflow creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Optimistic lock or overlap conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictErrorResponse"
...
  /bookings/{bookingId}/check-in:
    post:
      summary: Check in to booking
      parameters:
        - in: path
          name: bookingId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "200":
          description: Check-in recorded
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: RBAC denied or suspended account policy
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    CreateBookingRequest:
      type: object
      required:
        [facilityId, bookingDate, startTime, endTime, purpose, attendees]
      properties:
        facilityId:
          type: string
          format: uuid
        bookingDate:
          type: string
          format: date
        startTime:
          type: string
          pattern: '^([01]\\d|2[0-3]):[0-5]\\d$'
        endTime:
          type: string
          pattern: '^([01]\\d|2[0-3]):[0-5]\\d$'
        purpose:
          type: string
          minLength: 3
        attendees:
          type: integer
          minimum: 1
        bookedForUserId:
          type: string
          format: uuid
        recurrenceRule:
          type: string
    BookingResponse:
      type: object
      required: [id, status, facility, startDateTime, endDateTime, version]
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED]
        facility:
          $ref: "#/components/schemas/FacilityResponse"
        startDateTime:
          type: string
          format: date-time
        endDateTime:
          type: string
          format: date-time
        version:
          type: integer
    FacilityResponse:
      type: object
      required: [id, name, type, capacity, location, building, floor, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: "#/components/schemas/FacilityType"
        capacity:
          type: integer
        location:
          type: string
        building:
          type: string
        floor:
          type: string
        status:
          type: string
          enum: [ACTIVE, OUT_OF_SERVICE]
...
    CheckInRequest:
      type: object
      required: [method]
      properties:
        method:
          type: string
          enum: [QR, MANUAL]
...
    ErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time
    ConflictErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            currentVersion:
              type: integer
            conflictingBookingId:
              type: string
              format: uuid
...
    FacilityType:
      type: string
      enum:
        [
          LECTURE_HALL,
          LAB,
          MEETING_ROOM,
          AUDITORIUM,
          EQUIPMENT,
          SPORTS_FACILITY,
        ]
...